requirements.txt
aiohttp==3.9.5
python-dotenv==1.0.1
tqdm==4.66.4
Pillow==10.3.0
aiofiles==23.2.1